Rename id_model/id_graph to val_graph
