+++ csrc/id_model/to_string.h
@@ -7,8 +7,8 @@
 // clang-format on
 #pragma once
 
-#include <id_model/id_graph.h>
 #include <ir/all_nodes.h>
+#include <val_graph.h>
 
 #include <string>
 #include <vector>
