+++ csrc/id_model/validation_utils.h
@@ -8,8 +8,8 @@
 #pragma once
 
 #include <compute_at_map.h>
-#include <id_model/id_graph.h>
 #include <id_model/id_model.h>
+#include <val_graph.h>
 
 namespace nvfuser {
