+++ csrc/val_graph.cpp
@@ -5,9 +5,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 // clang-format on
-#include <id_model/id_graph.h>
 #include <id_model/to_string.h>
 #include <ir/utils.h>
+#include <val_graph.h>
 
 namespace nvfuser {
